server/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
celery
django-celery-results
django-cors-headers
drf-yasg
Pillow
django-storages
boto3
stripe
requests
